Update the bs.game package substantially.
This focused around the addition of the Reaction interface, which
describes things that happen after a given move, allowing moves and
their reactions (which are often other moves) to essentially form a
linked list containing the events of a single round of play.
Also, Moves were expanded to reference the players that made them,
and a Player class was created accordingly. Players and Moves work
together to keep track of the changing distribution of cards across
the table.