Expand Player class to track player order
Each Player now keeps track of a successor, the immediate next player
in the playing sequence. In addition to retrieving successors of a
player, it is also possible to retrieve the first successor that is
still playing (has not yet won), because players also now keep track
of whether or not they have won. To facilitate this and make the
successor relation more usable, Player now implements Iterable, with
a player's elements of iteration being itself, its successor, its
successor's successor, and so on, until reaching the original player.