Add CallResult constructor that accepts revealed
The contents of a Play are not always known when it is created. The
new constructor accepts a separate list of cards describing the
contents of a previously-unknown play, along with that play and a
call, and then processes the call as if the list of cards had been
returned by a call to play.getCards().