Implement new, more accurate lastMove method
Because it skipped over players who were out of the game, the
previous implementation failed to properly return the most recent
move in situations where the last move was a successful call by a
player with no hand, who then won. The new implementation corrects
this by storing a reference to the next round in CallResult objects,
allowing the entire move sequence to be traversed until the end is
reached.